back_end/requirements.txt
blinker==1.9.0
click==8.3.0
colorama==0.4.6
Flask==3.1.2
flask-cors==6.0.1
Flask-SQLAlchemy==3.1.1
greenlet==3.2.4
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.3
PyMySQL==1.1.2
SQLAlchemy==2.0.44
typing_extensions==4.15.0
Werkzeug==3.1.3
locust>=2.32.0
redis>=5.0
python-dotenv>=1.0
